+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-#v1.0.0..
+#v1.0.0 - 2014-01-02
   * FEATURE - Optional per-mailbox override of default recording hash
   * FEATURE - Optional rerecording of voicemail messages
   * FEATURE - I18n numeric methods